make sure splash start and end markers have correct position
Compiler can change position of start/end splash markers or even optimize-out
them completely, because they are not used in the fw. 
Make them part of the splash data block, so their possition is always correct.
util/xbm2lbm was changed together with other c "images" to make it consitent.


git-svn-id: https://open9x.googlecode.com/svn/trunk@188 aa04c113-223f-6118-1b62-88c24e5be702