Major patch to split haptic and audio queues into separate systems.
Patch is testing and working on the sky9xboard and the stock board.  I dont have a gruvin board to test on.

Currently only events that trigger haptic alerts are WARNING1-3 and TRIM MIDDLE.  

Next step is for bertrand to use the hooks to enable a per alert option to choose enable haptic.

Also... translations are added but the menu file does not use the translation yet - as I have not got my head round how this works!

git-svn-id: https://open9x.googlecode.com/svn/trunk@534 aa04c113-223f-6118-1b62-88c24e5be702